+++ chickenEggs/objects/Game/OrdinaryChicken.java
@@ -5,7 +5,7 @@ import static chickenEggs.interfaces.variables.*;
 import java.util.ArrayList;
 
 public class OrdinaryChicken extends Chicken {
-    public OrdinaryChicken (int x, int y, int w, int h, float xmingame, float xmaxgame, float ymingame, float ymaxgame) {
+    public OrdinaryChicken (int x, int y, int w, int h, int xmingame, int xmaxgame, int ymingame, int ymaxgame) {
         super(x, y, w, h , chicken[0] , xmingame, xmaxgame, ymingame, ymaxgame);
         this.health = 1;
         this.eggDuration = 100;
