+++ chickenEggs/objects/Game/SuperChicken.java
@@ -4,7 +4,7 @@ import static chickenEggs.interfaces.variables.*;
 
 import java.util.ArrayList;
 public class SuperChicken extends Chicken {
-    public SuperChicken (int x, int y, int w, int h, float xmingame, float xmaxgame, float ymingame, float ymaxgame) {
+    public SuperChicken (int x, int y, int w, int h, int xmingame, int xmaxgame, int ymingame, int ymaxgame) {
         super(x, y, w, h , chicken[2] , xmingame, xmaxgame, ymingame, ymaxgame);
         this.health = 3;
         this.eggDuration = 60;
