+++ chickenEggs/objects/Game/UltimateChicken.java
@@ -4,7 +4,7 @@ import static chickenEggs.interfaces.variables.*;
 
 import java.util.ArrayList;
 public class UltimateChicken extends Chicken {
-    public UltimateChicken(int x, int y, int w, int h, float xmingame, float xmaxgame, float ymingame, float ymaxgame) {
+    public UltimateChicken(int x, int y, int w, int h, int xmingame, int xmaxgame, int ymingame, int ymaxgame) {
         super(x, y, w, h , chicken[3] , xmingame, xmaxgame, ymingame, ymaxgame);
         this.health = 4;
         this.eggDuration = 40;
